requirements.txt
asyncio
numpy
opencv-python==4.4.0.42
websockets
requests
tensorflow==2.3.1
Keras==2.4.3
pillow
matplotlib
pandas
progressbar2
wandb
black